[ALPS04714118] wlan: Add timestamp to data stall event
Due to some data stall event is time critical,
so extend event id field to 32 bytes to include
time stamp after event id.

MTK-Commit-Id: 576f9de328c8c1d5544577156cd5578c01798337

CR-Id: ALPS04714118
Feature: [Module]Wi-Fi Driver

+++ common/wlan_lib.c
@@ -10268,19 +10268,19 @@ void wlanCustomMonitorFunction(struct ADAPTER *prAdapter,
 			prAdapter->rWifiVar.u4TxLowRateThreshole)
 			KAL_REPORT_ERROR_EVENT(prAdapter,
 				EVENT_TX_LOW_RATE,
-				(uint16_t)sizeof(u_int8_t),
+				(uint16_t)sizeof(uint32_t),
 				FALSE);
 		else if (prLinkQualityInfo->u4CurRxRate <
 			prAdapter->rWifiVar.u4RxLowRateThreshole)
 			KAL_REPORT_ERROR_EVENT(prAdapter,
 				EVENT_RX_LOW_RATE,
-				(uint16_t)sizeof(u_int8_t),
+				(uint16_t)sizeof(uint32_t),
 				FALSE);
 		else if (prLinkQualityInfo->u4CurTxPer >
 			prAdapter->rWifiVar.u4PerHighThreshole)
 			KAL_REPORT_ERROR_EVENT(prAdapter,
 				EVENT_PER_HIGH,
-				(uint16_t)sizeof(u_int8_t),
+				(uint16_t)sizeof(uint32_t),
 				FALSE);
 	}
 }
